solana: initiate swap (exact in) (#41)
* some clean-up

* solana: initiate swap works

* solana: add token interface for 2022 support

* solana: add tests; reorganize

* solana: more token interface

* solana: remove console.log

* solana: and another one

---------
