+++ solana/programs/swap-layer/src/composite/mod.rs
@@ -4,12 +4,13 @@ use crate::{
     error::SwapLayerError,
     state::{Custodian, Peer},
     utils::{
+        self,
         jupiter_v6::{self, cpi::SharedAccountsRouteArgs, JUPITER_V6_PROGRAM_ID},
         AnchorInstructionData,
     },
 };
 use anchor_lang::prelude::*;
-use anchor_spl::{associated_token, token};
+use anchor_spl::{associated_token, token, token_interface};
 use common::{
     admin::utils::{
         assistant::{self, only_authorized},
@@ -345,16 +346,17 @@ pub struct CompleteSwap<'info> {
         associated_token::mint = dst_mint,
         associated_token::authority = authority
     )]
-    pub dst_swap_token: Box<Account<'info, token::TokenAccount>>,
+    pub dst_swap_token: Box<InterfaceAccount<'info, token_interface::TokenAccount>>,
 
     /// This account must be verified as the source mint for the swap.
     pub usdc: Usdc<'info>,
 
     /// This account must be verified as the destination mint for the swap.
     #[account(constraint = usdc.key() != dst_mint.key() @ SwapLayerError::SameMint)]
-    pub dst_mint: Box<Account<'info, token::Mint>>,
+    pub dst_mint: Box<InterfaceAccount<'info, token_interface::Mint>>,
 
     pub token_program: Program<'info, token::Token>,
+    pub dst_token_program: Interface<'info, token_interface::TokenInterface>,
     associated_token_program: Program<'info, associated_token::AssociatedToken>,
     system_program: Program<'info, System>,
 }
@@ -398,7 +400,7 @@ impl<'info> CompleteSwap<'info> {
         ))?;
 
         token::close_account(CpiContext::new_with_signer(
-            self.token_program.to_account_info(),
+            self.dst_token_program.to_account_info(),
             token::CloseAccount {
                 account: self.dst_swap_token.to_account_info(),
                 destination,
@@ -520,12 +522,15 @@ impl<'info> JupiterV6SharedAccountsRoute<'info> {
         Ok((accounts, args, cpi_account_infos.to_vec()))
     }
 
-    pub fn invoke_cpi(
+    pub fn swap_exact_in(
         &self,
         args: SharedAccountsRouteArgs,
         signer_seeds: &[&[u8]],
         cpi_remaining_accounts: Vec<AccountInfo<'info>>,
-    ) -> Result<()> {
+        limit_amount: Option<u64>,
+    ) -> Result<u64> {
+        let limit_amount = limit_amount.unwrap_or(utils::jupiter_v6::compute_min_amount_out(&args));
+
         jupiter_v6::cpi::shared_accounts_route(
             CpiContext::new_with_signer(
                 self.jupiter_v6_program.to_account_info(),
@@ -548,6 +553,18 @@ impl<'info> JupiterV6SharedAccountsRoute<'info> {
             )
             .with_remaining_accounts(cpi_remaining_accounts),
             args,
+        )?;
+
+        // After the swap, we reload the destination token account to get the correct amount.
+        let amount_out = token::TokenAccount::try_deserialize_unchecked(
+            &mut &self.dst_custody_token.data.borrow()[..],
         )
+        .map(|token| token.amount)?;
+
+        // Rarely do I use the gte macro, but this is a good use case for it. I want to display the
+        // amounts if the limit amount is not met.
+        require_gte!(amount_out, limit_amount, SwapLayerError::SwapFailed);
+
+        Ok(amount_out)
     }
 }
